+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # unreleased
 * Deprecate `GrantedPaymentInstrumentUpdate` and add `GrantorUpdatedGrantedPaymentMethod` and `RecipientUpdatedGrantedPaymentMethod`
+* Add account_type support for transaction sale, verification, and payment_method create/update
 
 # 2.94.0
 * Add `bin` field to `payment_method_nonce_details`